refactor dao
