+++ newbank/server/data/passwords.txt
@@ -1,4 +1,3 @@
 bhagy:1234
 christina:1234
 john:1234
-andy:1234
